+++ attributes/default.rb
@@ -20,14 +20,13 @@
 # default recipe
 # set location of backup file to restore from
 default['mcs']['restore']['file'] = ''
-# user and organization for Chef-managed server
+
+# managed organization for Chef-managed server
 default['mcs']['org']['name'] = 'chef_managed_org'
 default['mcs']['org']['full_name'] = 'Chef Managed Organization'
-default['mcs']['managed_user']['dir'] = '/etc/opscode/managed'
-default['mcs']['managed_user']['user_name'] = 'chef_managed_user'
-default['mcs']['managed_user']['first_name'] = 'Chef'
-default['mcs']['managed_user']['last_name'] = 'Managed'
+# if you want an email address for the managed organization users
 default['mcs']['managed_user']['email'] = 'you@example.com'
+# if you want a non-random password for the user
 default['mcs']['managed_user']['password'] = nil
 
 # backup recipe